requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
kaggle==1.6.3
PyGithub==2.1.1
loguru==0.7.2
aiohttp==3.9.1
pydantic==2.5.2
